Revert "Feature chat"
