--- lib/core/constants/asset_images.dart
@@ -1,3 +1,0 @@
-class AssetImages{
-  static const logo= "images/whats.png";
-}
