--- lib/core/shared_widgets/custom_button.dart
@@ -1,64 +1,0 @@
-import 'package:chat_app/core/themes/colors_app.dart';
-import 'package:chat_app/core/themes/styles.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-class CustomButton extends StatelessWidget{
-  const CustomButton({
-    super.key,
-    this.backgroundColor,
-    this.borderRadius,
-    required this.text,
-    this.fontSized,
-    this.textColor,
-    required this.func,
-    this.width,
-    this.height=43,
-    this.isLoading = false,
-    //this.svgPicture,
-  });
-final double? width;
-final double? height;
-final Color? backgroundColor;
-final Color? textColor;
-final BorderRadius? borderRadius;
-final String text;
-final double? fontSized;
-final VoidCallback func;
-final bool isLoading;
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: SizedBox(
-        width: double.infinity,
-        child: ElevatedButton(
-          onPressed: func,
-          style: ElevatedButton.styleFrom(
-            backgroundColor: ColorsApp.primaryColor,
-            padding: EdgeInsets.symmetric(horizontal: 48,vertical: 12),
-            shape: RoundedRectangleBorder(
-              borderRadius: borderRadius ?? BorderRadius.circular(24),
-            )
-          ),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Center(
-                child: isLoading
-                ? const CircularProgressIndicator()
-                    :Text(
-                  text,
-                  style: styles.textStyle24.copyWith(
-                    color: textColor ?? Colors.black,
-                  ),
-                )
-              )
-            ],
-          ),
-
-        ),
-      ),
-    );
-  }
-
-}
