--- lib/core/themes/colors_app.dart
@@ -1,8 +1,0 @@
-import 'dart:ui';
-
-abstract class ColorsApp{
-  static const primaryColor = Color(0xff4cb050);
-  static const secondColor = Color(0xff000000);
-  static const thirdColor = Color(0xff9e9e9e);
-  static const fourthColor = Color(0xffffffff);
-}
