--- lib/core/themes/styles.dart
@@ -1,18 +1,0 @@
-import 'dart:ui';
-
-import 'package:flutter/material.dart';
-
-abstract class styles{
-  static TextStyle textStyle24 =TextStyle(
-    fontSize: 24,
-    color: Colors.black,
-    fontWeight: FontWeight.bold,
-  );
-
-  static TextStyle textStyle16 =TextStyle(
-    fontSize: 16,
-    color: Colors.black,
-    fontWeight: FontWeight.w700,
-  );
-
-}
