+++ lib/features/auth/views/widgets/login_body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-class StatusPage extends StatelessWidget{
+class LoginBody extends StatelessWidget{
   @override
   Widget build(BuildContext context) {
     return Scaffold();
