--- lib/features/call/presentation/views/call_screen.dart
@@ -1,10 +1,0 @@
-import 'package:chat_app/features/call/presentation/views/widgets/call_page.dart';
-import 'package:flutter/cupertino.dart';
-
-class CallScreen extends StatelessWidget{
-  @override
-  Widget build(BuildContext context) {
-     return CallPage();
-  }
-
-}
