--- lib/features/call/presentation/views/widgets/call_page.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-class CallPage extends StatelessWidget{
-  @override
-  Widget build(BuildContext context) {
-     return Scaffold();
-  }
-
-}
