--- lib/features/status/presentation/views/status_screen.dart
@@ -1,10 +1,0 @@
-import 'package:chat_app/features/status/presentation/views/widgets/status_page.dart';
-import 'package:flutter/cupertino.dart';
-
-class StatusScreen extends StatelessWidget{
-  @override
-  Widget build(BuildContext context) {
-     return StatusPage();
-  }
-
-}
