+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,8 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import path_provider_foundation
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  PathProviderPlugin.register(with: registry.registrar(forPlugin: "PathProviderPlugin"))
 }
