+++ pubspec.yaml
@@ -35,8 +35,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
-  google_fonts: ^6.2.1
-  fluttertoast: ^8.2.8
 
 dev_dependencies:
   flutter_test:
